Configure service support email address for environments
The support email inbox is not considered secret so should be configured
within `appsettings.json` for each environment.

However, only feedback/support in production should go to the Regional
Services inbox, and GOV.UK Notify won't send emails to anyone not on the
developer list when using a dev API key, so for lower environments this
has been configured with an appropriate email address that meets these
constraints.